Update floattidf.c
Hi,
once you have floated the ID, makes sense to ensure your back-end by tapping into a popular protocol and proposing a back-end fix to prevent the problem recurring 
and to ensure it is you and no one else to take possession of the associated private/public key function. Just make a persuasive case and let the wisdom of the crowd decide!